+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Common & useful classes, resources, extensions. Based on Symfony framework.
 17. Documentation > Development > update
 18. Docker > rename `php-cli` service to `php`
 19. Remove `Version` class (Value Object > Version of software). Use `Version` class from `meritoo/common-library` package.
+20. Base Dependency Injection (DI) extension, the `BaseExtension` class
 
 # 0.0.1